refactor: allow mutation of `getTheme` return object; abstract `cloneDeep` and `isObject`
